Add testing for and fix bugs in tokenize
Add a test file for the process module, and add a complete test suite (class)
for the tokenize function. Fix the bugs found and break the regex into parts
for clarity. Also move test files into a tests directory, and rename the source
directory to source_lyrics to avoid confusion.